requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-genai==1.26.0
openai==1.51.0
pydantic==2.5.0
requests==2.31.0
